This moves the serialization logic to Elastic.CommonSchema
The serilog formatter can just call ecsEvent.Serialize() and so can any
other integration we want to write while making sure they all write the
same.

Wrote a custom formatter for Base that writes the events in the order
that we want and deconstructs log.level from log.

We can ilmerge Utf8Json or go the route we did with the client and ship
with it internally