Add trace context fields in ElasticsearchLogger (#115)
This commit adds trace context fields to log events
when using ElasticsearchLogger.

The trace context fields are extracted from the current
System.Diagnostics.Activity if not null. When the Id
format used is W3C, the Activity.TraceId and Activity.SpanId
will be used to set Trace.Id and Span.Id, respectively. For
any other id format, Trace.Id and Span.Id will be set from
Activity.RootId and Activity.Id, respectively. When there
is no current Activity but CorrelationManager has an
ActivityId, this will be used for the Trace.Id.

Co-authored-by: Russ Cam <russ.cam@elastic.co>