Fixes #86 only write log level once
With the new serialization generation we were emitting log level twice
once under `log.level` and once as `log: { level: }`

This PR addresses this to ensure we never emit the latter but are able
to read both versions of the json.

This to make sure the JSON we emit is compatible with the
`json.expand_keys` option in filebeat as identified as @elan-slovelock

This PR also makes sure we never emit `log: {}` if only log level is
set.