Fix #162 only enrich with APM if enabled
Since 1.7 the Elastic Agent has an `Enabled` configuration that allows
users to disable the agent through configuration.

The Serilog and NLOG enrichers are now aware of this setting and will no
longer inject empty tracing properties if the agent is not configured to
collect them.

This bumps the minimum supported version of the Elastic.Apm agent to
1.8.