[backport] AOT Annotations and CI smoke tests for AOT publishes (#530)
Adds AOT annotations for all published projects and ensures they are
advertised as AOT compatible.


Create a smoke test binary that writes to an in memory elasticsearch
endpoint for our NLog and Extensions loggers.

Serilog is currently failing with:

```
ecs-aot-smoketest failed with 2 error(s) (58.2s) → examples/ecs-aot-smoketest/bin/release/net9.0/osx-arm64/ecs-aot-smoketest.dll
    ILC : Trim analysis error IL2067: Serilog.Capturing.PropertyValueConverter.TryConvertStructure(Object,Type,Destructuring,StructureValue&): 'type' argument does not satisfy 'DynamicallyAccessedMemberTypes.PublicProperties' in call to 'Serilog.Capturing.PropertyValueConverter.CreateStructureValue(Object,Type,Boolean)'. The parameter 'type' of method 'Serilog.Capturing.PropertyValueConverter.TryConvertStructure(Object,Type,Destructuring,StructureValue&)' does not have matching annotations. The source value must declare at least the same requirements as those declared on the target location it is assigned to.
    /Users/mpdreamz/.nuget/packages/microsoft.dotnet.ilcompiler/9.0.7/build/Microsoft.NETCore.Native.targets(317,5): error MSB3073: The command ""/Users/mpdreamz/.nuget/packages/runtime.osx-arm64.microsoft.dotnet.ilcompiler/9.0.7/tools/ilc" @"obj/release/net9.0/osx-arm64/native/ecs-aot-smoketest.ilc.rsp"" exited with code -1.
```

Will create an issue with Serilog to see if it can be resolved.


This removes https://github.com/benaadams/Ben.Demystifier as its not AOT
compatible. AOT compatibility beats nicer stacktrace strings as a
feature to keep supporting IMO.

This bumps NLog to 6.0.0 to take benefit of its work to be AOT
compatible (cc @snakefoot)

Also be aware wuth this backport we bump NLog to 6.0.0 for our `8.x`
releases.

We typically do not require major version bumps in our deps to force a
bump in our major version. cc @snakefoot